Refactor prefetch generation out of discovery service.  Add support for CRMI extensions.
