+++ hapi-fhir-server-cds-hooks/src/test/java/ca/uhn/hapi/fhir/cdshooks/svc/cr/discovery/CrDiscoveryServiceR4Test.java
@@ -40,5 +40,4 @@ public class CrDiscoveryServiceR4Test extends BaseCrTest {
 			"}";
 		assertEquals(expected, actual);
 	}
-
 }
